put volunteers.py, manageLabor.py and participants.py in the same file and combined the function checkUserParticipant()
