--- app/models/eventLabor.py
@@ -1,8 +1,0 @@
-from app.models import *
-from app.models.user import User
-from app.models.event import Event
-
-class EventLabor(baseModel):
-    user = ForeignKeyField(User, null = False)
-    event = ForeignKeyField(Event, null = False)
-    didWork = BooleanField(null = False, default = False)
